refactored version
