+++ dosma/__init__.py
@@ -1,9 +1,8 @@
-"""The core module contains functions and classes for musculoskeletal analysis.
+"""The core module contains functions and classes for medical image analysis.
 """
-from dosma.core.io.dicom_io import *  # noqa
 from dosma.core import *  # noqa
-from dosma.core.io.nifti_io import *  # noqa
-from dosma.utils.device import Device  # noqa
+from dosma.utils.collect_env import collect_env_info  # noqa
+from dosma.utils.env import debug  # noqa
 
 # This line will be programatically read/write by setup.py.
 # Leave them at the bottom of this file and don't touch them.
