+++ dosma/core/__init__.py
@@ -1,14 +1,26 @@
-from dosma.core import (  # noqa: F401
-    io,
+from dosma.core import io
+
+from dosma.core import (
+    device,
+    fitting,
     med_volume,
+    numpy_routines,
     orientation,
+    quant_vals,
+    registration,
 )
 
-from dosma.core.io import *
+from dosma.core.device import *  # noqa
+from dosma.core.fitting import *  # noqa
+from dosma.core.io import *  # noqa
 from dosma.core.med_volume import *  # noqa
 from dosma.core.orientation import *  # noqa
+from dosma.core.registration import *  # noqa
 
-__all__ = []
+__all__ = ["numpy_routines", "quant_vals"]
+__all__.extend(device.__all__)
+__all__.extend(fitting.__all__)
 __all__.extend(io.__all__)
 __all__.extend(med_volume.__all__)
 __all__.extend(orientation.__all__)
+__all__.extend(registration.__all__)
