+++ dosma/core/io/__init__.py
@@ -5,12 +5,12 @@ from dosma.core.io import (  # noqa: F401
     nifti_io,
 )
 from dosma.core.io.dicom_io import *  # noqa
-from dosma.core.io.format_io import *  # noqa
+from dosma.core.io.format_io import ImageDataFormat  # noqa
 from dosma.core.io.format_io_utils import *  # noqa
 from dosma.core.io.nifti_io import *  # noqa
 
 __all__ = []
 __all__.extend(dicom_io.__all__)
-__all__.extend(format_io.__all__)
+__all__.extend(["ImageDataFormat"])
 __all__.extend(format_io_utils.__all__)
 __all__.extend(nifti_io.__all__)
