+++ dosma/core/registration.py
@@ -15,11 +15,11 @@ from tqdm import tqdm
 from tqdm.contrib.concurrent import process_map
 
 from dosma import file_constants as fc
+from dosma.core.device import cpu_device
 from dosma.core.io.nifti_io import NiftiReader, NiftiWriter
 from dosma.core.med_volume import MedicalVolume
 from dosma.defaults import preferences
 from dosma.utils import env
-from dosma.utils.device import cpu_device
 
 __all__ = ["register", "apply_warp", "symlink_elastix", "unlink_elastix"]
