+++ dosma/scan_sequences/__init__.py
@@ -2,7 +2,7 @@ from dosma.scan_sequences import (  # noqa: F401
     mri,
     scans,
 )
-from dosma.scan_sequences.mri import *
+from dosma.scan_sequences.mri import *  # noqa
 from dosma.scan_sequences.scans import *  # noqa
 
 __all__ = []
