+++ tests/utils/test_registration.py
@@ -9,9 +9,9 @@ import numpy as np
 import dosma.file_constants as fc
 from dosma.core.io.dicom_io import DicomReader
 from dosma.core.io.format_io import ImageDataFormat
+from dosma.core.registration import apply_warp, register
 from dosma.scan_sequences import CubeQuant, QDess
 from dosma.utils import env
-from dosma.utils.registration import apply_warp, register
 
 from .. import util
